init: Force a log to the kernel log for boot animation complete
Need to force a log statement to the kernel log to determine when the
boot animation is completed.  Boot animation exit is when the UI is ready
for display to the user.

If the logcat output is redirected to kmsg there is a 6 second addition to
overall boot time.  This is obviously not acceptable.

There is always boot chart but that requires a rebuild and a special know how on
how to do that so that should be relugated to developers.

So that leaves us with being able to replicate this from just a kernel log without
adding the overhead of special tools or code.

+++ init/signal_handler.c
@@ -63,6 +63,10 @@ static int wait_for_one_process(int block)
 
     NOTICE("process '%s', pid %d exited\n", svc->name, pid);
 
+    /* Strictly for logging boot animation exit to the kernel */
+    if (!strcmp("bootanim", svc->name))
+         ERROR("Boot Animation exit\n");
+
     if (!(svc->flags & SVC_ONESHOT)) {
         kill(-pid, SIGKILL);
         NOTICE("process '%s' killing any children in process group\n", svc->name);
